Moved the tabor driver with the new logic into their own package "awgs_new_driver"
Renamed the old_base and old_tabor back to their original names base and tabor
