+++ qupulse/_program/seqc.py
@@ -35,7 +35,7 @@ from qupulse._program.waveforms import Waveform
 from qupulse._program._loop import Loop
 
 from qupulse._program.volatile import VolatileRepetitionCount, VolatileProperty
-from qupulse.hardware.awgs.old_base import ProgramEntry
+from qupulse.hardware.awgs.base import ProgramEntry
 
 try:
     import zhinst.utils
