+++ tests/hardware/awg_base_tests.py
@@ -1,11 +1,11 @@
-from typing import Callable, Iterable, List, Optional, Set, Tuple
+from typing import Callable, Iterable, Optional, Set, Tuple
 import unittest
 import warnings
 
 from qupulse import ChannelID
 from qupulse._program._loop import Loop
-from qupulse.hardware.awgs.base import AWGDevice, AWGChannel, AWGChannelTuple, AWGMarkerChannel
-from qupulse.hardware.awgs.features import ChannelSynchronization, ProgramManagement, VoltageRange, \
+from qupulse.hardware.awgs_new_driver.base import AWGDevice, AWGChannel, AWGChannelTuple, AWGMarkerChannel
+from qupulse.hardware.awgs_new_driver.features import ChannelSynchronization, ProgramManagement, VoltageRange, \
     AmplitudeOffsetHandling
 from qupulse.utils.types import Collection
