+++ tests/hardware/old_tabor_simulator_based_tests.py
@@ -7,7 +7,7 @@ import os
 import pytabor
 import numpy as np
 
-from qupulse.hardware.awgs.old_tabor import TaborAWGRepresentation, TaborChannelPair
+from qupulse.hardware.awgs.tabor import TaborAWGRepresentation, TaborChannelPair
 from qupulse._program.tabor import TaborSegment, PlottableProgram, TaborException, TableDescription, TableEntry
 from typing import List, Tuple, Optional, Any
